[OCT-384] feat: add speaker back (#408)
* feat: fix some bug and add speakers in db

* feat : fix the add of speaker on the back and link the front

* feat: adding waitingscreen components

* feat : add tests backend

* feat: fix some bug and add speakers in db

* feat : fix the add of speaker on the back and link the front

* add public event/track/user in back

* fix: fix some bug for event and track edit

* remove log

* remove wainting screen of mailys

* make test better and change some controleur behaviour

* feat: add 2 tests files for event usecase and make tests cleaner, change a little bit of behavious

* fix testsuit

* fix use case and some test

* fix update event

* fix : publih event

* fix : close event and event behaviour

* fix: get private event by id

* fix: jsute remove useless argument in api.js

* fix: some track error

+++ backend/src/events/domain/usecases/deleteEvent.usecase.ts
@@ -1,14 +1,28 @@
-import { Inject } from '@nestjs/common';
+import {
+  Inject,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { IEventGateway } from '../gateways/events.gateway';
 import { EventObject } from '../event';
+import { IUserGateway } from 'src/users/domain/gateways/users.gateway';
 
 export class DeleteEventUsecase {
   constructor(
     @Inject('EventGateway')
     private eventGateway: IEventGateway,
+    @Inject('UserGateway')
+    private userGateway: IUserGateway,
   ) {}
 
-  async execute(eventId: string): Promise<EventObject> {
+  async execute(eventId: string, currentEmail: string): Promise<EventObject> {
+    const currentUser = await this.userGateway.getUserByEmail(currentEmail);
+    if (!currentUser) throw new NotFoundException('User not foud');
+    const event = await this.eventGateway.getEventById(eventId);
+    if (!event) throw new NotFoundException('Event not Found');
+    if (!event.canBeEditBy(currentUser))
+      throw new UnauthorizedException("Current user can't delete this event");
+
     return await this.eventGateway.deleteEvent(eventId);
   }
 }
